hyperv: Move hv_connection_id to hyperv-tlfs.h
This structure should be in hyperv-tlfs.h anyway, since it is part of
the TLFS document.
The definition conflicts with one added in hvgdk.h as part of the mshv
driver so must be moved to hyperv-tlfs.h.

Signed-off-by: Nuno Das Neves <nunodasneves@linux.microsoft.com>
(cherry picked from commit b7ccb352278356bbe86a08ff6fd4337e610192a8)