Disable CONFIG_X86_MPPARSE with MSHV_VTL
When CONFIG_X86_MPPARSE is enabled, the kernel will scan low memory
looking for MP tables. We definitely don't want this to happen, because
low memory is controlled by VTL0 and may contain actual valid tables
during a servicing reboot.

This change disables CONFIG_X86_MPPARSE for the Underhill kernel. To do
this, it has to make X86_MPPARSE selectable when OF is enabled--normally
it is only selectable when ACPI is enabled and otherwise not, and we do
not select ACPI... so we get X86_MPPARSE automatically.

This change also updates MSHV_VTL to explicitly depend on having some
features disabled so that we don't accidentally enable them in the
future. These are the features where enabling them would cause security
or data loss bugs.

Signed-off-by: Saurabh Sengar <ssengar@linux.microsoft.com>